web-scrape-cdk/requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.0.0
boto3
pytest
pytest-mock
